Change UserModel's id field's type from long to int

+++ src/main/java/com/example/model/UserModel.java
@@ -6,12 +6,12 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class UserModel {
-    private long id;
+    private int id;
     private String userName;
     private String email;
     private long passwordHash;
 
-    public UserModel(long id, String userName, String email, long passwordHash) {
+    public UserModel(int id, String userName, String email, long passwordHash) {
         this.id = id;
         this.userName = userName;
         this.email = email;
